k-classification/requirements.txt
mysql_connector_repackaged==0.3.1
nltk==3.8.1
numpy==1.23.3
pandas==1.5.0
scikit_learn==1.2.1
scipy==1.10.1
textblob==0.17.1
transformers==4.24.0
vaderSentiment==3.3.2
